[Refactor] : .yml change
