+++ src/main/java/com/bipa/bizsurvey/domain/admin/api/AdminController.java
@@ -5,8 +5,8 @@ import com.bipa.bizsurvey.domain.admin.application.AdminUserService;
 import com.bipa.bizsurvey.domain.admin.dto.claim.ClaimDetailResponse;
 import com.bipa.bizsurvey.domain.admin.dto.claim.ClaimListResponse;
 import com.bipa.bizsurvey.domain.admin.dto.user.UserSearchRequest;
-import com.bipa.bizsurvey.domain.community.service.PostService;
-import com.bipa.bizsurvey.domain.community.service.SurveyPostService;
+import com.bipa.bizsurvey.domain.community.application.PostService;
+import com.bipa.bizsurvey.domain.community.application.SurveyPostService;
 import com.bipa.bizsurvey.domain.user.application.UserService;
 import com.bipa.bizsurvey.domain.user.dto.mypage.UserInfoResponse;
 import com.bipa.bizsurvey.domain.user.enums.Plan;
